// PetRadar-Web-UI/src/app/api/petradar/model/userPetUpdateModel.ts
/**
 * PetRadar.Web.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PetSizeEnum } from './petSizeEnum';
import { PetSpeciesEnum } from './petSpeciesEnum';
import { PetSexEnum } from './petSexEnum';


export interface UserPetUpdateModel { 
    name?: string | null;
    species?: PetSpeciesEnum;
    breed?: string | null;
    color?: string | null;
    sex?: PetSexEnum;
    size?: PetSizeEnum;
    birthDate?: string | null;
    approximateAge?: number | null;
    weight?: number | null;
    description?: string | null;
    isNeutered?: boolean | null;
    allergies?: string | null;
    medicalNotes?: string | null;
}
export namespace UserPetUpdateModel {
}