// PetRadar-Web-UI/src/app/api/petradar/model/userPetViewModel.ts
/**
 * PetRadar.Web.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserPetViewModel { 
    id?: number;
    userId?: number;
    name?: string | null;
    species?: string | null;
    breed?: string | null;
    color?: string | null;
    sex?: string | null;
    size?: string | null;
    birthDate?: string | null;
    approximateAge?: number | null;
    weight?: number | null;
    description?: string | null;
    photoURL?: string | null;
    additionalPhotosURL?: string | null;
    isNeutered?: boolean | null;
    allergies?: string | null;
    medicalNotes?: string | null;
}